[#20] Add game index page, and add question index to answering page
