+++ src/pages/Game/Questions.less
@@ -2,13 +2,13 @@
   padding: 30px 20px 0 !important;
   max-width: 800px;
   font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-  background: url("../../images/game/background/hill-1.svg"), url("../../images/game/background/hill-2.svg"), url("../../images/game/background/hill-3.svg");
+  background: url("../../../images/game/background/hill-1.svg"), url("../../../images/game/background/hill-2.svg"), url("../../../images/game/background/hill-3.svg");
   background-color: #dddddd !important;
   background-repeat: no-repeat;
   animation: animatedBackground 10s linear infinite;
   min-height: 100vh;
   .Game {
-    background: url("../../images/game/background/panda.svg") no-repeat 2% 89%;
+    background: url("../../../images/game/background/panda.svg") no-repeat 2% 89%;
 
     .app-header {
       display: inline-block;
@@ -62,11 +62,11 @@
     }
 
     .option input:checked + b:before {
-      content: url("../../images/game/yes.svg");
+      content: url("../../../images/game/yes.svg");
     }
 
     .option input:not(:checked) + b:before {
-      content: url("../../images/game/No.svg");
+      content: url("../../../images/game/No.svg");
     }
 
     .background-placeholder {
