+++ src/pages/Game/index.js
@@ -1,2 +1,3 @@
 export Question from '../../components/Question'
 export Result from './Result'
+export Questions from './Questions'
